pyproject.toml
[project]
name = "llm-mbti-test"
version = "0.1.0"
description = "Run MBTI personality tests on LLMs via OpenMBTI MCP server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "httpx[socks]>=0.27.0",
]

[project.scripts]
llm-mbti = "main:main"
